+++ usage/usage.go
@@ -18,7 +18,6 @@ import (
 	"time"
 
 	"github.com/essentialkaos/ek/v13/fmtc"
-	"github.com/essentialkaos/ek/v13/mathutil"
 	"github.com/essentialkaos/ek/v13/strutil"
 	"github.com/essentialkaos/ek/v13/version"
 )
@@ -793,7 +792,7 @@ func getMaxCommandSize(commands []*Command, group string) int {
 			continue
 		}
 
-		size = mathutil.Max(size, getCommandSize(command)+2)
+		size = max(size, getCommandSize(command)+2)
 	}
 
 	return size
@@ -804,7 +803,7 @@ func getMaxOptionSize(options []*Option) int {
 	var size int
 
 	for _, option := range options {
-		size = mathutil.Max(size, getOptionSize(option)+2)
+		size = max(size, getOptionSize(option)+2)
 	}
 
 	return size
@@ -815,7 +814,7 @@ func getMaxEnvVarSize(envVars []*Env) int {
 	var size int
 
 	for _, env := range envVars {
-		size = mathutil.Max(size, strutil.Len(env.Name)+2)
+		size = max(size, strutil.Len(env.Name)+2)
 	}
 
 	return size
@@ -919,7 +918,7 @@ func printNewVersionInfo(curVersion, newVersion string, releaseDate time.Time) {
 
 // wrapText wraps text to the specified length with indentation
 func wrapText(text string, indent, maxLen int) string {
-	size := mathutil.Max(_DEFAULT_WRAP_LEN, maxLen) - indent
+	size := max(_DEFAULT_WRAP_LEN, maxLen) - indent
 
 	if strutil.LenVisual(fmtc.Clean(text)) <= size {
 		return text